A few notes
Got image showing for slider bar

Had tried something without committing and rolled back with undos. Was trying to move JS out of html.erb to .js so could use Chrome to see what was going on.

Have made some changes to test opacity which is working but not well enough for my use.