Saving to extent_json working.
Now need to draw with that.

And eventually get rid of extent_array since it's not working.

street_extent_json was street_extend_json for id.

Had street_extent_json disabled: true, and this didn't allow data to be written to db (thought maybe it didn't just allow writing on the webpage)