Added a blank paragraph on some static pages
The heading was going under the navbar for reasons I couldn't understand. Some other pages had a paragraph for notices at the top of the page and they worked, so did the same for the other pages

The navbar has 20px margin. The top elements of the other pages have zero padding and margins.