Duplicate working except:
The segment isn't drawn on the map. This may be true for edit too.

Had to add @maps to dup (just like new)

Have to figure out how to add to index, although not essential. In some ways better w/o because it adds to the line length and going to show then dup isn't that much work.