View and creating for map references
https://stackoverflow.com/questions/14146023/align-checkboxes-for-f-collection-check-boxes-with-simple-form#14146400

Edit/New view is OK, but not saving. Nothing to tell it where to save.

`bootstrap: {check_inline: true}` instead of `:item_wrapper_class => 'inline'` renders the same. Not in numerical order as it was before.

has_and_belongs_to_many is not generally preferred but if don't need access to the join is OK

http://api.rubyonrails.org/classes/ActiveRecord/Associations/ClassMethods.html#method-i-has_and_belongs_to_many

`@maps` from streets>edit instead of `Map.all` as edit now includes sort by year

https://github.com/comfy/comfy-bootstrap-form